Refactoriser la LiveResultTable en fonction de ses composants Fixes #236
